+++ docker/php/fpm/bin/symfony/03-configure.sh
@@ -1,0 +1,9 @@
+#!/bin/sh
+set -e
+
+# opcache
+echo "Configure opcache..."
+echo "opcache.preload = /srv/app/config/preload.php" >> /usr/local/etc/php/conf.d/docker-php.ini
+
+# done
+echo "Done."