+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make: Targets for `debug` and `info`.
 - PHP: `docker-entrypoint`
 - Docs: Create an example for Symfony based applications
+- PHP: Scripts for `create`, `install` and `configure` the Symfony framework
 
 ### Changed
 
@@ -37,12 +38,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PHP: Upgrade to v8.1.13
 - Web: Upgrade to v1.23.2
 - Move `opcache` setting to general `php.ini`
+- PHP: Copy Symfony scripts in the container
 
 ### Deprecated
 
 ### Removed
 
 - Unused regex for `.env.dist` in `.dockerignore`
+- Remove `opcache.preload` from `php.ini`
 
 ### Fixed